LayoutTests/http/tests/security/cross-frame-access-port-explicit-domain.html

<html>
<head>
    <script src="resources/cross-frame-access.js"></script>
</head>
<body>
<p>This test currently fails because we check the port and protocol even if document.domain is explicitly set (rdar://problem/5366437).</p> 
<iframe id="aFrame" name="aFrame"></iframe>
<pre id="console"></pre>
<script>
    // Explicitly set the domain.
    document.domain = "127.0.0.1";

    var url = "http://127.0.0.1:8080/security/resources/cross-frame-iframe-with-explicit-domain-set.html";
    var iframeId ="aFrame";
    var passMessage = "PASS: Cross frame access to a different port, after explicitly setting document.domain, was allowed!";
    var failMessage = "Fail: Cross frame access to a different port, after explicitly setting document.domain, was denied.";
    canAccessFrame(url, iframeId, passMessage, failMessage);
</script>
</body>
</html>
