LayoutTests/http/tests/security/cross-frame-access-port-explicit-domain-expected.txt

CONSOLE MESSAGE: line 1: Unsafe JavaScript attempt to access frame with URL http://127.0.0.1:8000/security/resources/cross-frame-iframe.html from frame with URL http://127.0.0.1:8000/security/cross-frame-access-parent-explicit-domain.html. Domains, protocols and ports must match.

This test currently fails because we check the port and protocol even if document.domain is explicitly set (rdar://problem/5366437).


PASS: Cross frame access to a different port, after explicitly setting document.domain, was allowed!


--------
Frame: 'aFrame'
--------
PASS: Cross frame access to a different port, after explicitly setting document.domain, was allowed!

Inner iframe with explicit document.domain set.
